Introduce stream/trampoline-based reification
This commit introduces a trampoline-style approach to reification.  It's essentially
inspired by the miniKanren stream processing alongside which this library is
primarily employed.  The main difference is that this implementation employs
the `Generator.send` method instead of callables (i.e. goals in miniKanren).

It removes reliance on the Python stack and, as a result, is much more scalable.
Also, it provides a mechanism through which substructures can be traversed but
not reconstructed.

Additions/extensions to the reification functions must follow the `yield`
paradigm when recursively calling `_reify` and precede object construction
with `yield construction_sentinel` to signify that substructure reification
is occurring next (i.e. so that the caller can prevent it, if desired).

Closes #13 and closes #10.